+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - factory reset: optional restore wifi settings
 - direct methods: return with error in case of file system failures
 - fix bug in main loop when channel is closed by sender
+- report cargo package version
 
 ## [0.4.1] Q2 2022
 - general: